refactor: utils extraction.
