+++ src/guide/region.js
@@ -1,5 +1,5 @@
 const Util = require('../util');
-const PathUtil = require('./path-util');
+const PathUtil = require('./util/path');
 const Guide = require('./base');
 
 class Region extends Guide {
